+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## unreleased
 - Early cutoff based on size difference for `Make.*.get_distance`.
+- Bit fiddling optimizations: `snoc_ones`, `snoc_zeros`.
 
 ## 0.1.0